agent/kubernetes_bot/pyproject.toml
[project]
name = "kubernetes-bot"
version = "0.1.0"
description = "Kubernetes management agent using ADK"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "google-adk[a2a]>=1.13.0",
    "google-genai>=1.17.0",
    "python-dotenv>=1.1.0",
    "kubernetes>=29.0.0",
    "uvicorn>=0.34.2",
    "starlette>=0.46.2",
    "sse-starlette>=2.3.5",
]
